Fold the migration handover event helpers into activities.go
Two functions did not need a file of their own. They sit next to
checkHandoverOnce, which is what feeds them.

Renames the test file to match, since handover_events.go no longer exists in
this package.
